+++ lib/imports/imports.dart
@@ -1,16 +1,17 @@
 //------------------------PACKAGES----------------------------------------------
-export 'package:flutter/material.dart';
 export 'dart:convert';
+export 'package:flutter/material.dart';
 export 'package:flutter_screenutil/flutter_screenutil.dart';
 
 //-------------------------VIEWS------------------------------------------------
 export 'package:curanetwork_library/utils/theme/AppTheme.dart';
 
 //------------------------UTILS-------------------------------------------------
+export 'package:curanetwork_library/utils/tools.dart';
+export 'package:curanetwork_library/utils/constants.dart';
+export 'package:shared_preferences/shared_preferences.dart';
 export 'package:curanetwork_library/utils/values/params.dart';
 export 'package:curanetwork_library/utils/animations/animation.dart';
-export 'package:curanetwork_library/utils/constants.dart';
-export 'package:curanetwork_library/utils/tools.dart';
 
 
 //------------------------Services----------------------------------------------